requirements.txt
folium==0.14.0
networkx==3.1
numpy==1.24.3
openrouteservice==2.3.3
geopy==2.3.0
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3
requests==2.31.0 